refactor(explorer): extract route target resolution from collector
Move protocol detection and service target PSI resolution into ArmeriaRouteTargetExtractor to shrink ArmeriaRouteCollector.
